requirements.txt
chromadb
sentence-transformers
groq
flask
flask-cors
PyMuPDF
Pillow
pytesseract